+++ servers/view/index.php
@@ -13,7 +13,7 @@ require("../../header.php");
         <div class="col-md-3">
 
           <div class="box box-danger">
-            <div class="box-body box-profile" id="1stDiv" style="display: none;">
+            <div class="box-body box-profile" id="div-1" style="display: none;">
               <div class="profile-user-img" style="border: 0;"><i class="fa fa-server fa-5x"></i></div>
 
               <h3 class="profile-username text-center" id="serverName"></h3>
@@ -30,11 +30,23 @@ require("../../header.php");
 
           </div>
 
+          <div class="box box-danger">
+            <div class="box-header with-border">
+              <h3 class="box-title">Broadcast message</h3>
+            </div>
+            <div class="box-body" id="div-2" style="display: none;">
+              <textarea id="txtBroadcastMessage" class="form-control"></textarea><br />
+              <button id="btnBroadcastMessage" class="btn btn-primary form-control">Send broadcast</button>
+            </div>
+            <div class="loading" id="loader-1"></div>
+
+          </div>
+
         </div>
 
         <div class="col-md-9">
           <div class="box box-danger">
-            <div class="box-body" id="2ndDiv" style="display: none;">
+            <div class="box-body" id="div-3" style="display: none;">
               <div class="row">
                 <div class="col-sm-9 border-right">
                   <div class="description-block">
@@ -64,6 +76,32 @@ require("../../header.php");
               <div class="row">
                 <div class="col-sm-9 border-right">
                   <div class="description-block">
+                    <h5 class="description-header">Is home server</h5>
+                    <span class="description-text" id="serverHomeServer"></span>
+                  </div>
+                </div>
+                <div class="col-sm-3">
+                  <div class="description-block">
+                    <a id="btnChangeHomeServer" class="btn btn-primary form-control">Change home server</a>
+                  </div>
+                </div>
+              </div>
+              <div class="row">
+                <div class="col-sm-9 border-right">
+                  <div class="description-block">
+                    <h5 class="description-header">Home server kingdom</h5>
+                    <span class="description-text" id="serverKingdom"></span>
+                  </div>
+                </div>
+                <div class="col-sm-3">
+                  <div class="description-block">
+                    <a id="btnChangeServerKingdom" class="btn btn-primary form-control">Change server kingdom</a>
+                  </div>
+                </div>
+              </div>
+              <div class="row">
+                <div class="col-sm-9 border-right">
+                  <div class="description-block">
                     <h5 class="description-header">Wurm time</h5>
                     <span class="description-text" id="serverWurmTime"></span>
                   </div>
@@ -74,27 +112,26 @@ require("../../header.php");
                   </div>
                 </div>
               </div>
-
-
               <div class="row">
                 <div class="col-sm-9 border-right">
                   <div class="description-block">
-                    <h5 class="description-header">Broadcast Message</h5>
-                    <input type="text" class="form-control" id="txtBroadcastMessage" />
+                    <h5 class="description-header">Player limit</h5>
+                    <span class="description-text" id="serverPlayerLimit"></span>
                   </div>
                 </div>
                 <div class="col-sm-3">
                   <div class="description-block">
-                    <a id="btnBroadcastMessage" class="btn btn-primary form-control">Send broadcast</a>
+                    <a id="btnChangePlayerLimit" class="btn btn-primary form-control">Change player limit</a>
                   </div>
                 </div>
               </div>
+
             </div>
-            <div class="box-body" id="loader-1">
+            <div class="box-body" id="loader-2">
               <div class="loading"></div>
             </div>
           </div>
-          
+
         </div>
       </div>
     </section>
@@ -128,34 +165,6 @@ require("../../header.php");
       </div>
     </div>
   </div>
-  <div class="modal" id="modalMute" tabindex="-1" role="dialog">
-    <div class="modal-dialog">
-      <div class="modal-content">
-        <div class="modal-header">
-          <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-            <span aria-hidden="true">×</span></button>
-          <h4 class="modal-title">Mute player</h4>
-        </div>
-        <div class="modal-body" id="modalMuteLoader" style="display:none;"><div class="loading"></div></div>
-        <form role="form" id="formMutePlayer">
-          <div class="modal-body">
-            <div class="form-group">
-              <label>How many hours?</label>
-              <input type="number" class="form-control" id="txtMuteHours" placeholder="Enter how many hours to mute player" />
-            </div>
-            <div class="form-group">
-              <label>Reason</label>
-              <input type="text" class="form-control" id="txtMuteReason" placeholder="Reason for mute" />
-            </div>
-          </div>
-          <div class="modal-footer">
-            <button type="button" class="btn btn-default pull-left" data-dismiss="modal">Cancel</button>
-            <button type="submit" class="btn btn-danger">Mute!</button>
-          </div>
-        </form>
-      </div>
-    </div>
-  </div>
 
   <input type="hidden" id="txtServerID" value="<?php echo $_GET['id']; ?>" />
 
@@ -182,7 +191,6 @@ require("../../header.php");
               }
             }
             else if(response.success) {
-              console.log(response);
               /**
                * Left col
                */
@@ -192,6 +200,7 @@ require("../../header.php");
 
               if(response.COUNT.success == false) {
                 $('#btnShutDown').prop('disabled', true);
+                $('#btnBroadcastMessage').prop('disabled', true);
               }
 
               /**
@@ -211,6 +220,36 @@ require("../../header.php");
                 $('#serverCluster').html('FREEDOM');
               }
 
+              if(response.HOMESERVER == 1) {
+                $('#serverHomeServer').html('Yes');
+              }
+              else {
+                $('#serverHomeServer').html('No');
+              }
+
+              switch(response.KINGDOM) {
+                case '0':
+                  $('#serverKingdom').html('No kingdom');
+                  break;
+                case '1':
+                  $('#serverKingdom').html('Jenn-Kellon');
+                  break;
+                case '2':
+                  $('#serverKingdom').html('Mol Rehan');
+                  break;
+                case '3':
+                  $('#serverKingdom').html('Horde of the Summoned');
+                  break;
+                case '4':
+                  $('#serverKingdom').html('Freedom Isles');
+                  break;
+                default:
+                  $('#serverKingdom').html('Unknown kingdom');
+                  break;
+              }
+
+              $('#serverPlayerLimit').html(response.MAXPLAYERS);
+
               $('#serverWurmTime').html((response.WURMTIME.success == false) ? 'Offline' : response.WURMTIME);
 
             }
@@ -218,26 +257,32 @@ require("../../header.php");
               swal("Error!", "Could not load this server", "error");
             }
 
-            $('#1stDiv').show();
-            $('#2ndDiv').show();
+            $('#div-1').show();
+            $('#div-2').show();
+            $('#div-3').show();
             $('#loader-0').hide();
             $('#loader-1').hide();
+            $('#loader-2').hide();
 
           },
           error: function(error) {
             console.log(error);
             swal("Failed", "It looks like we couldn't proccess your request at this time. Please try again later.", "error");
-            $('#userList').show();
-            $('#loader').hide();
+            $('#div-1').show();
+            $('#div-2').show();
+            $('#div-3').show();
+            $('#loader-0').hide();
+            $('#loader-1').hide();
+            $('#loader-2').hide();
           }
         });
-      
+
       }
 
       $('#btnShutDown').on('click', function(e) {
         e.preventDefault();
         $('#modalShutdown').modal('show');
-      
+
       });
 
       $('#formShutdown').on('submit', function(e) {
@@ -323,6 +368,7 @@ require("../../header.php");
 
             $('#btnBroadcastMessage').prop('disabled', false);
             $('#btnBroadcastMessage').html('Send broadcast');
+            $('#txtBroadcastMessage').val('');
 
           },
           error: function(error) {
@@ -332,7 +378,336 @@ require("../../header.php");
             $('#btnBroadcastMessage').html('Send broadcast');
           }
         });
-      
+
+      });
+
+      $('#btnChangeGameMode').on('click', function(e) {
+        e.preventDefault();
+        swal({
+          title: 'Change game mode',
+          text: '<select id="txtGameMode" class="form-control"><option value="0">PVE</option><option value="1">PVP</option></select>',
+          html: true,
+          showCancelButton: true,
+          showConfirmButton: true,
+          confirmButtonText: 'Change',
+          cancelButtonText: 'Cancel',
+          showLoaderOnConfirm: true,
+          closeOnConfirm: false
+        },
+        function(isConfirm) {
+          if(isConfirm) {
+            var newGameMode = $('#txtGameMode').val();
+
+            $.ajax({
+              type: 'POST',
+              url: 'process.php',
+              data: {doing: "changeGameMode", newGameMode: newGameMode, serverID: serverID},
+              dataType: 'json',
+              success: function(response) {
+                console.log(response);
+                if(response.error) {
+                  switch(response.error.message) {
+                    case 'Missing database':
+                      swal("Missing Databases", "Couldn't find the server database. Please double check your config file.", "error");
+                      break;
+                    default:
+                      swal("Error", response.error.message, "error");
+                      break;
+                  }
+                }
+                else if(response.success) {
+                  var txtGameMode = "";
+                  switch(newGameMode) {
+                    case '0':
+                      txtGameMode = 'PVE';
+                      break;
+                    case '1':
+                      txtGameMode = 'PVP';
+                      break;
+                    default:
+                      txtGameMode = 'Unknown game mode';
+                      break;
+                  }
+                  swal('Game mode changed!', 'The game mode for this server has been changed to [ ' + txtGameMode + ' ]!', 'success');
+                  $('#serverGameMode').html(txtGameMode);
+                }
+                else {
+                  swal('Failed to change!', 'We could not proccess this request at this time.', 'error');
+                }
+
+              },
+              error: function(error) {
+                console.log(error);
+                swal('Failed', 'It looks like we couldn\'t proccess your request at this time. Please try again later.', 'error');
+              }
+            });
+          }
+        });
+
+      });
+
+      $('#btnChangeCluster').on('click', function(e) {
+        e.preventDefault();
+        swal({
+          title: 'Change game cluster',
+          text: '<select id="txtGameCluster" class="form-control"><option value="0">Freedom</option><option value="1">EPIC</option></select>',
+          html: true,
+          showCancelButton: true,
+          showConfirmButton: true,
+          confirmButtonText: 'Change',
+          cancelButtonText: 'Cancel',
+          showLoaderOnConfirm: true,
+          closeOnConfirm: false
+        },
+        function(isConfirm) {
+          if(isConfirm) {
+            var newGameCluster = $('#txtGameCluster').val();
+
+            $.ajax({
+              type: 'POST',
+              url: 'process.php',
+              data: {doing: "changeGameCluster", newGameCluster: newGameCluster, serverID: serverID},
+              dataType: 'json',
+              success: function(response) {
+                console.log(response);
+                if(response.error) {
+                  switch(response.error.message) {
+                    case 'Missing database':
+                      swal("Missing Databases", "Couldn't find the server database. Please double check your config file.", "error");
+                      break;
+                    default:
+                      swal("Error", response.error.message, "error");
+                      break;
+                  }
+                }
+                else if(response.success) {
+                  var txtGameCluster = "";
+                  switch(newGameCluster) {
+                    case '0':
+                      txtGameCluster = 'Freedom';
+                      break;
+                    case '1':
+                      txtGameCluster = 'Epic';
+                      break;
+                    default:
+                      txtGameCluster = 'Unknown game cluster';
+                      break;
+                  }
+                  swal('Game cluster changed!', 'The game cluster for this server has been changed to [ ' + txtGameCluster + ' ]!', 'success');
+                  $('#serverCluster').html(txtGameCluster);
+                }
+                else {
+                  swal('Failed to change!', 'We could not proccess this request at this time.', 'error');
+                }
+
+              },
+              error: function(error) {
+                console.log(error);
+                swal('Failed', 'It looks like we couldn\'t proccess your request at this time. Please try again later.', 'error');
+              }
+            });
+          }
+        });
+
+      });
+
+      $('#btnChangeHomeServer').on('click', function(e) {
+        e.preventDefault();
+        swal({
+          title: 'Is home server?',
+          text: '<select id="txtHomeServer" class="form-control"><option value="1">Yes</option><option value="0">No</option></select>',
+          html: true,
+          showCancelButton: true,
+          showConfirmButton: true,
+          confirmButtonText: 'Change',
+          cancelButtonText: 'Cancel',
+          showLoaderOnConfirm: true,
+          closeOnConfirm: false
+        },
+        function(isConfirm) {
+          if(isConfirm) {
+            var homeServer = $('#txtHomeServer').val();
+
+            $.ajax({
+              type: 'POST',
+              url: 'process.php',
+              data: {doing: "changeHomeServer", homeServer: homeServer, serverID: serverID},
+              dataType: 'json',
+              success: function(response) {
+                console.log(response);
+                if(response.error) {
+                  switch(response.error.message) {
+                    case 'Missing database':
+                      swal("Missing Databases", "Couldn't find the server database. Please double check your config file.", "error");
+                      break;
+                    default:
+                      swal("Error", response.error.message, "error");
+                      break;
+                  }
+                }
+                else if(response.success) {
+                  var txtHomeServer = "";
+                  switch(homeServer) {
+                    case '0':
+                      txtHomeServer = 'No';
+                      break;
+                    case '1':
+                      txtHomeServer = 'Yes';
+                      break;
+                    default:
+                      txtHomeServer = 'Unknown home server setting';
+                      break;
+                  }
+                  swal('Home server changed!', 'The home server setting for this server has been changed to [ ' + txtHomeServer + ' ]!', 'success');
+                  $('#serverHomeServer').html(txtHomeServer);
+                }
+                else {
+                  swal('Failed to change!', 'We could not proccess this request at this time.', 'error');
+                }
+
+              },
+              error: function(error) {
+                console.log(error);
+                swal('Failed', 'It looks like we couldn\'t proccess your request at this time. Please try again later.', 'error');
+              }
+            });
+          }
+        });
+
+      });
+
+      $('#btnChangeServerKingdom').on('click', function(e) {
+        e.preventDefault();
+        swal({
+          title: 'Change home server kingdom',
+          text: '<select id="txtHomeServerKingdom" class="form-control"><option value="0">No kingdom</option><option value="1">Jenn-Kellon</option><option value="2">Mol Rehan</option><option value="3">Horde of the Summoned</option><option value="4">Freedom Isles</option></select>',
+          html: true,
+          showCancelButton: true,
+          showConfirmButton: true,
+          confirmButtonText: 'Change',
+          cancelButtonText: 'Cancel',
+          showLoaderOnConfirm: true,
+          closeOnConfirm: false
+        },
+        function(isConfirm) {
+          if(isConfirm) {
+            var homeServerKingdom = $('#txtHomeServerKingdom').val();
+
+            $.ajax({
+              type: 'POST',
+              url: 'process.php',
+              data: {doing: "changeHomeServerKingdom", homeServerKingdom: homeServerKingdom, serverID: serverID},
+              dataType: 'json',
+              success: function(response) {
+                console.log(response);
+                if(response.error) {
+                  switch(response.error.message) {
+                    case 'Missing database':
+                      swal("Missing Databases", "Couldn't find the server database. Please double check your config file.", "error");
+                      break;
+                    default:
+                      swal("Error", response.error.message, "error");
+                      break;
+                  }
+                }
+                else if(response.success) {
+                  var txtHomeServerKingdom = "";
+                  switch(homeServerKingdom) {
+                    case '0':
+                      $('#serverKingdom').html('No kingdom');
+                      break;
+                    case '1':
+                      $('#serverKingdom').html('Jenn-Kellon');
+                      break;
+                    case '2':
+                      $('#serverKingdom').html('Mol Rehan');
+                      break;
+                    case '3':
+                      $('#serverKingdom').html('Horde of the Summoned');
+                      break;
+                    case '4':
+                      $('#serverKingdom').html('Freedom Isles');
+                      break;
+                    default:
+                      $('#serverKingdom').html('Unknown kingdom');
+                      break;
+                  }
+                  swal('Home server kingdom changed!', 'The default kingdom for this server has been changed to [ ' + txtHomeServerKingdom + ' ]!', 'success');
+                  $('#serverKingdom').html(txtHomeServerKingdom);
+                }
+                else {
+                  swal('Failed to change!', 'We could not proccess this request at this time.', 'error');
+                }
+
+              },
+              error: function(error) {
+                console.log(error);
+                swal('Failed', 'It looks like we couldn\'t proccess your request at this time. Please try again later.', 'error');
+              }
+            });
+          }
+        });
+
+      });
+
+      $('#btnChangeWurmTime').on('click', function(e) {
+        e.preventDefault();
+        swal('Not implemented!', 'This function hasn\'t been implemented yet', 'error');
+      });
+
+      $('#btnChangePlayerLimit').on('click', function(e) {
+        e.preventDefault();
+        swal({
+          title: 'Change max players',
+          text: 'Change the max number of players that can be on this server',
+          type: 'input',
+          showCancelButton: true,
+          showConfirmButton: true,
+          confirmButtonText: 'Change',
+          cancelButtonText: 'Cancel',
+          showLoaderOnConfirm: true,
+          closeOnConfirm: false
+        },
+        function(inputValue) {
+          if(inputValue === false || inputValue === "" || inputValue == 0) {
+            swal.showInputError("You need to write a number and has to be greater than 0!");
+            return false;
+          }
+          else {
+            $.ajax({
+              type: 'POST',
+              url: 'process.php',
+              data: {doing: "changePlayerLimit", newPlayerLimit: inputValue, serverID: serverID},
+              dataType: 'json',
+              success: function(response) {
+                console.log(response);
+                if(response.error) {
+                  switch(response.error.message) {
+                    case 'Missing database':
+                      swal("Missing Databases", "Couldn't find the server database. Please double check your config file.", "error");
+                      break;
+                    default:
+                      swal("Error", response.error.message, "error");
+                      break;
+                  }
+                }
+                else if(response.success) {
+                  swal('Player limited changed!', 'The new player limit for this server has been changed to [ ' + inputValue + ' ]!', 'success');
+                  $('#serverPlayerLimit').html(inputValue);
+                }
+                else {
+                  swal('Failed to change!', 'We could not proccess this request at this time.', 'error');
+                }
+
+              },
+              error: function(error) {
+                console.log(error);
+                swal('Failed', 'It looks like we couldn\'t proccess your request at this time. Please try again later.', 'error');
+              }
+            });
+          }
+        });
+
       });
 
     });